Rewrite and changes (feedback)
Archer:
Removed static final variables to another class. This because of efficiency and to have one id variable keeping count.
Merged multiple methods and removed the now obsolete methods.
Did a rewrite on calculateWeightedScore to make it cleaner and more sensible.
Removed the saving of rounds, since they are not used anywhere.